requirements.txt
torch>=2.3.0,<2.4
transformers>=4.41.0,<5
peft>=0.11.0,<0.12
accelerate>=0.30.0,<1
PyYAML>=5.4
